+++ src/components/Form/career/CareerForm.jsx
@@ -4,15 +4,24 @@ import ResumeSection from '../../Resume/ResumeSection';
 import CareerFormItem from './CareerFormItem';
 import CareerEditForm from './CareerEditForm';
 import useCareerDetail from '@/hooks/useCareerDetail';
-import useCareerFormik from '@/hooks/useCareerFormik';
+import useCustomFormik from '@/hooks/useCustomFormik';
 
 const CareerForm = () => {
   const { id } = useParams();
   const [isNewForm, setIsNewForm] = useState(false);
   const { createCareer, careerList } = useCareerDetail(id);
 
-  const formik = useCareerFormik({
-    id,
+  const formik = useCustomFormik({
+    initialValues: {
+      resume_id: Number(id),
+      career_id: '',
+      company_name: '',
+      job_title: '',
+      career_status: '',
+      career_start_date: '',
+      career_end_date: '',
+      job_detail: '',
+    },
     onSubmitCallback: (values) => {
       createCareer.mutate(values);
       formik.resetForm();
